requirements.txt
alembic==1.13.2
beautifulsoup4==4.12.3
blinker==1.8.2
certifi==2024.7.4
cffi==1.17.0
charset-normalizer==3.3.2
click==8.1.7
cryptography==43.0.0
exchange_calendars==4.5.5
Flask==3.0.3
Flask-Login==0.6.3
Flask-Migrate==4.0.7
Flask-SQLAlchemy==3.1.1
frozendict==2.4.4
html5lib==1.1
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.4
jsonify==0.5
korean-lunar-calendar==0.3.1
lxml==5.2.2
Mako==1.3.5
MarkupSafe==2.1.5
multitasking==0.0.11
numpy==2.0.1
pandas==2.2.2
pandas_market_calendars==4.4.1
peewee==3.17.6
platformdirs==4.2.2
psycopg2-binary==2.9.9
pycparser==2.22
pyluach==2.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
requests==2.32.3
schwabdev==2.2.0
six==1.16.0
soupsieve==2.5
SQLAlchemy==2.0.31
toolz==0.12.1
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
webencodings==0.5.1
websockets==12.0
Werkzeug==3.0.3
yfinance==0.2.41
